:recycle: Chore: Move logic from base-client to common
