+++ .changeset/hungry-parents-fail.md
@@ -1,0 +1,5 @@
+---
+"@tevm/common": minor
+---
+
+Added logging and chainId params to common. This moves logic that was previously in createBaseClient to the common factory
