+++ packages/base-client/package.json
@@ -64,19 +64,18 @@
 	"dependencies": {
 		"@tevm/blockchain": "workspace:^",
 		"@tevm/common": "workspace:^",
-		"@tevm/errors": "workspace:1.1.0-next.45",
 		"@tevm/evm": "workspace:^",
 		"@tevm/jsonrpc": "workspace:^",
-		"@tevm/logger": "workspace:1.0.0-next.42",
+		"@tevm/logger": "workspace:^",
 		"@tevm/predeploys": "workspace:^",
 		"@tevm/state": "workspace:^",
-		"@tevm/sync-storage-persister": "workspace:1.1.0-next.45",
-		"@tevm/txpool": "workspace:1.1.0-next.45",
+		"@tevm/sync-storage-persister": "workspace:^",
+		"@tevm/txpool": "workspace:^",
 		"@tevm/utils": "workspace:^",
 		"@tevm/vm": "workspace:^"
 	},
 	"devDependencies": {
-		"@tevm/test-utils": "workspace:1.0.0-next.39",
+		"@tevm/test-utils": "workspace:^",
 		"@tevm/tsconfig": "workspace:^",
 		"@tevm/tsupconfig": "workspace:^"
 	},
