+++ packages/common/docs/functions/createCommon.md
@@ -6,7 +6,7 @@
 
 # Function: createCommon()
 
-> **createCommon**(`options`?): [`Common`](../classes/Common.md)
+> **createCommon**(`options`): [`Common`](../classes/Common.md)
 
 Creates an ethereumjs Common object used by the EVM
 to access chain and hardfork parameters and to provide
@@ -14,7 +14,7 @@ a unified and shared view on the network and hardfork state.
 
 ## Parameters
 
-▪ **options?**: [`CommonOptions`](../type-aliases/CommonOptions.md)= `{}`
+▪ **options**: [`CommonOptions`](../type-aliases/CommonOptions.md)
 
 ## Returns
