+++ packages/common/package.json
@@ -62,7 +62,8 @@
 		"typecheck": "tsc --noEmit"
 	},
 	"dependencies": {
-		"@ethereumjs/common": "^4.3.0"
+		"@ethereumjs/common": "^4.3.0",
+		"@tevm/logger": "workspace:1.0.0-next.42"
 	},
 	"devDependencies": {
 		"@tevm/tsconfig": "workspace:^",
