+++ packages/common/src/CommonOptions.ts
@@ -1,3 +1,4 @@
+import type { LogOptions } from '@tevm/logger'
 import type { Hardfork } from './Hardfork.js'
 
 /**
@@ -5,11 +6,19 @@ import type { Hardfork } from './Hardfork.js'
  */
 export type CommonOptions = {
 	/**
+	 * The network chain id
+	 */
+	chainId: bigint
+	/**
 	 * Hardfork to use. Defaults to `shanghai`
 	 */
-	hardfork?: Hardfork
+	hardfork: Hardfork
 	/**
 	 * Eips to enable. Defaults to `[1559, 4895]`
 	 */
 	eips?: ReadonlyArray<number>
+	/**
+	 * Tevm logger instance
+	 */
+	loggingLevel: LogOptions['level']
 }
